+++ src/client/src/pages/detail/detailPage.tsx
@@ -17,7 +17,6 @@ import { DetailPageContent } from "./detailPageContent.tsx";
 import { DetailSideNav } from "./detailSideNav.tsx";
 import { BoreholeFormInputs } from "./form/borehole/boreholePanelInterfaces.ts";
 import { LocationFormInputs, LocationFormSubmission } from "./form/location/locationPanelInterfaces.tsx";
-import { FormDirtyProvider } from "./formDirtyContext.tsx";
 import { useLabelingContext } from "./labeling/labelingInterfaces.tsx";
 import LabelingPanel from "./labeling/labelingPanel.tsx";
 import { SaveBar } from "./saveBar";
@@ -127,7 +126,7 @@ export const DetailPage: FC = () => {
     (location.pathname.endsWith("/borehole") && location.hash === "#general");
 
   return (
-    <FormDirtyProvider>
+    <>
       <DetailHeader borehole={borehole} editableByCurrentUser={editableByCurrentUser} triggerReset={triggerReset} />
       <LayoutBox>
         <SidebarBox>
@@ -169,6 +168,6 @@ export const DetailPage: FC = () => {
           {editingEnabled && shouldShowSaveBar && <SaveBar triggerSubmit={triggerSubmit} triggerReset={triggerReset} />}
         </Stack>
       </LayoutBox>
-    </FormDirtyProvider>
+    </>
   );
 };